# openapi.yaml
openapi: 3.0.0
info:
  title: Utdon API Documentation
  version: 1.1.0
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  responses:
    UnauthorizedError:
      description: Access token,user/password are missing, invalid, or session is not set
  schemas:
    Error:
      type: string
      description: Catched Error as string
...
security:
  - ApiKeyAuth: []
servers:
  - url: /api/v1
paths:
...
  /scrap/{url}:
    get:
      summary: Get provided url content
      description: Used to retrieve url content from the production server, github API tags...
      security:
        - ApiKeyAuth: []
      tags:
        - Core
      parameters:
        - in: path
          name: url
          required: true
          description: url to scrap http||https supported
          schema:
            type: string
      responses:
        "200":
          description: content as text
          content:
            application/text:
              schema:
                type: string
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...